+++ ft_varg.c
@@ -6,7 +6,7 @@
 /*   By: pclement <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/20 17:39:41 by pclement          #+#    #+#             */
-/*   Updated: 2017/12/20 20:02:23 by pclement         ###   ########.fr       */
+/*   Updated: 2017/12/21 13:32:14 by pclement         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,7 +101,7 @@ intmax_t	va_arg_intmax(va_list ap, char *v_type)
 uintmax_t	va_arg_uintmax(va_list ap, char *v_type)
 {
 	if (ft_strcmp(v_type, "size_t") == 0)
-		return ((unsigned short)va_arg(ap, uintmax_t));
+		return ((uintmax_t)va_arg(ap, uintmax_t));
 	if (ft_strcmp(v_type, "unsigned short int") == 0)
 		return ((unsigned short)va_arg(ap, uintmax_t));
 	if (ft_strcmp(v_type, "unsigned char") == 0)
